+++ lib/glimmer/swt/composite_proxy.rb
@@ -4,8 +4,6 @@ require 'glimmer/swt/widget_proxy'
 module Glimmer
   module SWT
     class CompositeProxy < WidgetProxy
-      attr_reader :layout
-      
       def initialize(parent, args, block)
         super(parent, args, block)
         @layout = default_layout
@@ -27,7 +25,23 @@ module Glimmer
       def layout=(the_layout)
         @layout = the_layout
       end
+      alias set_layout layout=
+      alias setLayout layout=
       
+      def get_layout
+        @layout
+      end
+      alias getLayout get_layout #TODO consider pregenerating these aliases with an easy method in the future
+
+      def pack(*args)
+        # No Op (just a shim) TODO consider if it should be implemented
+      end
+      
+      def layout​(changed = nil, all = nil)
+        # TODO implement layout(changed = nil, all = nil) just as per SWT API
+        @layout&.layout(self, changed)
+      end
+            
     end
     
   end