+++ lib/glimmer/util/proc_tracker.rb
@@ -1,5 +1,5 @@
 # Copyright (c) 2020-2021 Andy Maleh
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
 # "Software"), to deal in the Software without restriction, including
@@ -7,10 +7,10 @@
 # distribute, sublicense, and/or sell copies of the Software, and to
 # permit persons to whom the Software is furnished to do so, subject to
 # the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be
 # included in all copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 # MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -24,15 +24,24 @@ require 'delegate'
 module Glimmer
   module Util
     class ProcTracker < DelegateClass(Proc)
-      def initialize(proc)
-        super(proc)
+      attr_reader :owner
+      
+      def initialize(proc = nil, owner: nil, &block)
+        super(proc || block)
+        @owner = owner
       end
+      
       def call(*args)
         __getobj__.call(*args)
         @called = true
       end
+      
       def called?
         !!@called
+      end
+      
+      def respond_to?(method, *args, &block)
+        %w[owner called?].include?(method.to_s) || super(method, *args, &block)
       end
     end
   end